Address comments by adding separate providers
