+++ site/testsuites/testsuite-kotlin/src/test/kotlin/docs/tbdex/pfi/ExchangesApiProvider.kt
@@ -1,0 +1,21 @@
+package website.tbd.developer.site.docs.tbdex.pfi
+
+import website.tbd.developer.site.docs.utils.MockExchangesApiProvider
+import tbdex.sdk.protocol.models.Message
+import website.tbd.developer.site.docs.utils.*
+import tbdex.sdk.httpserver.models.*
+
+class ExchangesApiProvider: MockExchangesApiProvider() {
+    // :snippet-start: pfiOverviewWriteKt
+    fun write(message: Message) {
+        val data = mapOf(
+            "exchangeid" to message.metadata.exchangeId,
+            "messagekind" to message.metadata.kind,
+            "messageid" to message.metadata.id,
+            "subject" to message.metadata.from,
+            "message" to message.data
+        )
+        dataProvider.insert("exchange", data)
+    }
+    // :snippet-end:
+}
