+++ site/testsuites/testsuite-kotlin/src/test/kotlin/docs/tbdex/pfi/PfiStructureTest.kt
@@ -75,42 +75,3 @@ class PfiStructureTest {
         // :snippet-end:
     }
 }
-
-class ExchangesApiProvider: MockExchangesApiProvider() {
-    // :snippet-start: pfiOverviewWriteKt
-    fun write(message: Message) {
-        val data = mapOf(
-            "exchangeid" to message.metadata.exchangeId,
-            "messagekind" to message.metadata.kind,
-            "messageid" to message.metadata.id,
-            "subject" to message.metadata.from,
-            "message" to message.data
-        )
-        dataProvider.insert("exchange", data)
-    }
-    // :snippet-end:
-}
-
-class OfferingsApiProvider: MockOfferingsApiProvider() {
-
-    // :snippet-start: pfiOverviewReadOfferingsKt
-    override fun getOffering(id: String): Offering {
-        val result = dataProvider.get("offering", id ?: "")
-        return result as Offering
-    }
-
-    override fun getOfferings(filter: GetOfferingsFilter?): List<Offering> {
-        val results = dataProvider.query("offering", "*")
-        val offerings = mutableListOf<Offering>()
-
-        if (results == null) { return emptyList() }
-
-        for (result in results) {
-            val offering = Offering.parse(result as String)
-            offerings.add(offering)
-        }
-
-        return offerings
-    }
-    // :snippet-end:
-}
